added the password functionality to the helpers package. Adapter users package to helper.password
