+++ users/usergenerator.go
@@ -1,9 +1,9 @@
 package users
 
 import (
-	"github.com/goccy/go-json"
 	"fmt"
-	"runtime"
+
+	"github.com/goccy/go-json"
 
 	"github.com/pkg/errors"
 	"github.com/qlik-oss/gopherciser/enummap"
@@ -11,9 +11,6 @@ import (
 )
 
 type (
-	// Password
-	Password string
-
 	// Type type of scheduler
 	Type int
 	// Settings interface to be implemented by user generator
@@ -120,26 +117,6 @@ func NewUserGeneratorNone() UserGenerator {
 			GeneratorType: UserGeneratorNone,
 		},
 		&NoneUsers{},
-	}
-}
-
-// UnmarshalJSON unmarshal password from json
-func (passwd *Password) UnmarshalJSON(arg []byte) error {
-	var s string
-	if err := json.Unmarshal(arg, &s); err != nil {
-		return errors.Wrap(err, "failed to unmarshal password")
-	}
-	*passwd = Password(s)
-
-	return nil
-}
-
-//  MarshalJSON marshal password to json: replace with ***
-func (passwd *Password) MarshalJSON() ([]byte, error) {
-	if runtime.GOOS != "js" {
-		return []byte(`"***"`), nil
-	} else {
-		return []byte(fmt.Sprintf("\"%s\"", string(*passwd))), nil
 	}
 }
